Change package name, issue 5
