+++ src/main/java/com/telefonica/iot/perseo/Result.java
@@ -18,7 +18,7 @@
 * iot_support at tid dot es
 */
 
-package es.tid.fiware.perseo;
+package com.telefonica.iot.perseo;
 
 /**
  *
