+++ src/test/java/com/telefonica/iot/perseo/test/EventBeanMock.java
@@ -18,7 +18,7 @@
 * iot_support at tid dot es
 */
 
-package es.tid.fiware.perseo.test;
+package com.telefonica.iot.perseo.test;
 
 import com.espertech.esper.client.EventBean;
 import com.espertech.esper.client.EventPropertyDescriptor;
